web and model

+++ tests/test_models/test_engine/test_file_storage.py
@@ -19,10 +19,8 @@ class test_fileStorage(unittest.TestCase):
 
     def tearDown(self):
         """ Remove storage file at end of tests """
-        try:
+        if os.path.exists('file.json'):
             os.remove('file.json')
-        except:
-            pass
 
     def test_obj_list_empty(self):
         """ __objects is initially empty """
